feat: 프로필드랍다운 컴포넌트로 빼고 네비게이션 기능 추가
